[refactor] move protocol utils to common folder
