+++ src/common/BUILD.bazel
@@ -1,0 +1,10 @@
+cc_library(
+  name = "utils",
+  srcs = ["utils.cc"],
+  hdrs = ["utils.h"],
+  visibility = ["//visibility:public"],
+  deps = [
+    "@com_github_grpc_grpc//:grpc++",
+    "@com_google_protobuf//:protobuf_lite",
+  ],
+)
